requirements.txt
fastapi==0.104.0
uvicorn==0.23.2
SQLAlchemy==2.0.22
alembic==1.12.0
python-dotenv==1.0.0
aiosqlite==0.19.0
pydantic_settings==2.0.3
aiohttp==3.8.6